Implement CastNode in GPU, and more accurate FP16 training with master copy of FP32

+++ Source/ComputationNetworkLib/ComputationNode.cpp
@@ -1008,7 +1008,7 @@ void ComputationNode<ElemType>::WriteMinibatchWithFormatting(FILE* f,
                     double absSumLocal = 0;
                     for (size_t j = 0; j < jend; j++) // loop over elements
                     {
-                        absSumLocal += abs(seqData[i * istride + j * jstride]);
+                        absSumLocal += (double)abs(seqData[i * istride + j * jstride]);
                     }
                     absSum += absSumLocal;
                 }
